+++ x8086NetEmu/Adapters/Video/VideoAdapter.vb
@@ -57,7 +57,7 @@
     Private Memory(X8086.MemSize - 1) As Boolean
 
     Public Sub New(cpu As X8086)
-        cpu.TryAttachHook(New X8086.MemHandler(Function(address As UInt32, ByRef value As UInt32, mode As X8086.MemHookMode)
+        cpu.TryAttachHook(New X8086.MemHandler(Function(address As UInt32, ByRef value As UInt16, mode As X8086.MemHookMode)
                                                    If mode = X8086.MemHookMode.Write AndAlso
                                                             ((address >= mStartTextVideoAddress AndAlso address <= mEndTextVideoAddress) OrElse
                                                             (address >= mStartGraphicsVideoAddress AndAlso address <= mEndGraphicsVideoAddress)) Then
